[solo-node] move governance stuff to its own file.
