+++ runtime/solo/src/types.rs
@@ -14,29 +14,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod governance {
-	use crate::{CouncilCollective, TechnicalCommitteeInstance};
-	use ajuna_primitives::AccountId;
-	use frame_support::traits::EitherOfDiverse;
-	use frame_system::EnsureRoot;
-	use pallet_collective::{EnsureProportionAtLeast, EnsureProportionMoreThan};
-
-	pub(crate) type EnsureRootOrMoreThanHalfCouncil = EitherOfDiverse<
-		EnsureRoot<AccountId>,
-		EnsureProportionMoreThan<AccountId, CouncilCollective, 1, 2>,
-	>;
-
-	pub type EnsureRootOrMoreThanHalfTechnicalCommittee = EitherOfDiverse<
-		EnsureRoot<AccountId>,
-		EnsureProportionAtLeast<AccountId, TechnicalCommitteeInstance, 1, 2>,
-	>;
-
-	pub type EnsureRootOrAllTechnicalCommittee = EitherOfDiverse<
-		EnsureRoot<AccountId>,
-		EnsureProportionAtLeast<AccountId, TechnicalCommitteeInstance, 1, 1>,
-	>;
-}
-
 pub mod proxy {
 	use frame_support::{
 		pallet_prelude::MaxEncodedLen, traits::InstanceFilter, RuntimeDebugNoBound,
